Revert "fixed display drivers"
This reverts commit e19396a87cf43bf30a80fff975918a4c499663e2.

+++ Core/Inc/ili9486.h
@@ -8,15 +8,10 @@
 #ifndef INC_ILI9486_H_
 #define INC_ILI9486_H_
 
-#include <stdint.h>
-
-// Basic init
 void ili9486_init(void);
-
-// Core drawing routines
 void draw_pixel(uint16_t x, uint16_t y, uint16_t color);
 void fill_screen(uint16_t color);
-void lcd_printf(int x, int y, const char* str, uint16_t color);
+void lcd_printf(uint16_t x, uint16_t y, const char* str, uint16_t color);
